pyproject.toml
[project]
name = "app-builder"
version = "0.1.0"
description = "Multi-agent system for building software applications"
requires-python = ">=3.10"
authors = [
    {name = "App Builder Team"}
]

dependencies = [
    # Core framework - LangChain 1.0 and LangGraph 1.0
    "langchain>=1.1.0",           # LangChain 1.1.0 (Nov 24, 2025)
    "langchain-core>=1.1.0",      # LangChain Core 1.1.0 (Nov 21, 2025)
    "langgraph>=1.0.4",           # LangGraph 1.0.4 (Nov 25, 2025)
    "langgraph-checkpoint-postgres>=3.0.1",  # Compatible with LangGraph 1.0

    # MCP integration
    "langchain-mcp-adapters>=0.1.0",  # MCP support (March 2025)

    # LLM providers - All compatible with LangChain 1.0
    "langchain-google-genai>=3.2.0",   # Google GenAI 3.2.0 (Nov 24, 2025)
    "langchain-anthropic>=1.2.0",      # Anthropic 1.2.0 (Nov 24, 2025) - Optional
    "langchain-openai>=1.1.0",         # OpenAI 1.1.0 (Nov 24, 2025) - Optional

    # Database
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",

    # Testing
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",

    # Code quality
    "ruff>=0.1.0",
    "mypy>=1.7.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint]
ignore = ["E501"]  # Line too long (let formatter handle it)

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
